requirements.txt
certifi==2019.6.16
chardet==3.0.4
Django==2.2.4
djangorestframework==3.10.3
idna==2.8
psycopg2==2.8.3
pyflakes==2.1.1
python-dotenv==0.10.3
pytz==2019.1
requests==2.22.0
sqlparse==0.3.0
urllib3==1.25.3
